modified:   WebDevelopmentLab03/pages/DJ_Helper.py

+++ WebDevelopmentLab03/pages/DJ_Helper.py
@@ -70,7 +70,7 @@ with st.form("survey_form"):
                             similarSongList.append(i["name"])
                     elif boolExpressionArtist and boolExpressionPlays:
                         if float(i["artist"]["name"]) != inputArtist:
-                            if float(i["playcount"]) >= 100000.0:
+                            if int(i["playcount"]) >= 100000:
                                 similarSongList.append(i["name"])
                 else:
                     similarSongList.append(i["name"])
